requirements.txt
mysql-connector-python==9.0.0
netifaces==0.11.0
